# ebookstore-backend/src/main/java/com/ebookstore/dto/OrderItemResponse.java
package com.ebookstore.dto;

import java.math.BigDecimal;

public record OrderItemResponse(
        Long id,
        ProductSummaryResponse product,
        Integer quantity,
        BigDecimal unitPrice,
        BigDecimal subtotal
) {}