# openapi.yaml
openapi: 3.0.3
info:
  title: E-Bookstore API
  description: Complete REST API for the E-Bookstore capstone application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductSummaryResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        author: { type: string }
        price: { type: number, format: double }
        imageUrl: { type: string }
        rating: { type: number, format: double }
        categoryName: { type: string }
        brandName: { type: string }
        stockQuantity: { type: integer }
    ProductDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ProductSummaryResponse'
        - type: object
          properties:
            description: { type: string }
            isbn: { type: string }
            pages: { type: integer }
            language: { type: string }
            estimatedDeliveryDays: { type: integer }
            reviewCount: { type: integer }
            category:
              type: object
              properties:
                id: { type: integer, format: int64 }
                name: { type: string }
            brand:
              type: object
              properties:
                id: { type: integer, format: int64 }
                name: { type: string }
    ProductPageResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/ProductSummaryResponse' }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        currentPage: { type: integer }
        pageSize: { type: integer }
...
    PlaceOrderRequest:
      type: object
      required: [addressId, paymentMethod]
      properties:
        addressId: { type: integer, format: int64 }
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, GIFT_POINTS]
        useGiftPoints: { type: boolean, default: false }
        cardNumber: { type: string }
        cardExpiry: { type: string }
        cardCvv: { type: string }
...
    AddressResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        label: { type: string }
        street: { type: string }
        city: { type: string }
        state: { type: string }
        zipCode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
    OrderItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        product: { $ref: '#/components/schemas/ProductSummaryResponse' }
        quantity: { type: integer }
        unitPrice: { type: number, format: double }
        subtotal: { type: number, format: double }
    OrderSummaryResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { type: string }
        totalAmount: { type: number, format: double }
        placedAt: { type: string, format: date-time }
        itemCount: { type: integer }
        paymentMethod: { type: string }
        paymentStatus: { type: string }
    OrderDetailResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { type: string }
        totalAmount: { type: number, format: double }
        placedAt: { type: string, format: date-time }
        cancelledAt: { type: string, format: date-time }
        paymentMethod: { type: string }
        paymentStatus: { type: string }
        giftPointsUsed: { type: integer }
        giftPointsEarned: { type: integer }
        address: { $ref: '#/components/schemas/AddressResponse' }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItemResponse' }
    CreateReviewRequest:
      type: object
      required: [productId, rating]
      properties:
        productId: { type: integer, format: int64 }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
    ReviewResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        fullName: { type: string }
        rating: { type: integer }
        comment: { type: string }
        createdAt: { type: string, format: date-time }
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
paths:
...
  /api/products:
    get:
      tags: [Products]
      operationId: getProducts
      summary: List products with filters
      description: Returns a paginated list of products with optional filters for category, brand, price range, and search query.
      security: []
      parameters:
        - name: categoryId
          in: query
          schema: { type: integer, format: int64 }
        - name: brandId
          in: query
          schema: { type: integer, format: int64 }
        - name: search
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: sort
          in: query
          schema: { type: string, enum: [relevance, price_asc, price_desc, rating], default: relevance }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 12 }
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductPageResponse' }
...
  /api/products/{id}:
    get:
      tags: [Products]
      operationId: getProductById
      summary: Get product by ID
      description: Returns full product details including description, ISBN, pages, language, and review count.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductDetailResponse' }
        '404':
          description: Product not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/products/{id}/related:
    get:
      tags: [Products]
      operationId: getRelatedProducts
      summary: Get related products
      description: Returns up to 4 products in the same category, excluding the current product.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProductSummaryResponse' }
...
  /api/orders:
    get:
      tags: [Orders]
      operationId: getOrders
      summary: Get order history
      description: Returns paginated order history for the authenticated user, sorted by date descending.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Order history
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items: { $ref: '#/components/schemas/OrderSummaryResponse' }
                  totalElements: { type: integer }
                  totalPages: { type: integer }
                  currentPage: { type: integer }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      tags: [Orders]
      operationId: placeOrder
      summary: Place a new order
      description: Creates an order from the current cart. Validates stock, applies gift points discount, and clears the cart.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceOrderRequest' }
            example:
              addressId: 1
              paymentMethod: "CREDIT_CARD"
              useGiftPoints: false
              cardNumber: "4111111111111111"
              cardExpiry: "12/26"
              cardCvv: "123"
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDetailResponse' }
        '400':
          description: Empty cart or insufficient stock
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/orders/{id}:
    get:
      tags: [Orders]
      operationId: getOrderById
      summary: Get order by ID
      description: Returns full order details including items, address, and gift points.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDetailResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Order does not belong to user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/orders/{id}/cancel:
    put:
      tags: [Orders]
      operationId: cancelOrder
      summary: Cancel an order
      description: Cancels an order if it is in PENDING or CONFIRMED status and was placed within the last 48 hours. Restores stock and refunds gift points used.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDetailResponse' }
        '400':
          description: Order outside 48-hour window or wrong status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Order does not belong to user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/reviews:
    post:
      tags: [Reviews]
      operationId: createReview
      summary: Create a product review
      description: Submits a rating and comment for a product. Each user can review a product only once. Updates product average rating on creation.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReviewRequest' }
            example: { productId: 1, rating: 5, comment: "Excellent book!" }
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewResponse' }
        '400':
          description: Already reviewed this product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Product not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/reviews/product/{productId}:
    get:
      tags: [Reviews]
      operationId: getProductReviews
      summary: Get reviews for a product
      description: Returns all reviews for a product ordered by most recent first.
      security: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReviewResponse' }
        '404':
          description: Product not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...